refactor: rename new density test page
